+++ front/html/create-menu.html
@@ -1,4 +1,4 @@
-<div data-page="order" class="page">
+<div data-page="create-menu" class="page">
     <div class="navbar">
         <div class="navbar-inner">
             <div class="left"><a href="index.html" class="back link icon-only"><i class="icon icon-back"></i></a></div>
