+++ front/html/group-detail.html
@@ -1,4 +1,4 @@
-<div data-page="group" class="page">
+<div data-page="group-detail" class="page">
     <div class="navbar navbar-trans">
 
         <div class="navbar-inner  ">
@@ -12,7 +12,7 @@
             <a href="#tab-1" class="button button-fill button color-green toolbar-button">
                 查看菜单
             </a>
-            <a onclick="myApp.alert('加入成功!')" href="#tab-2" class="button button-fill color-red toolbar-button">
+            <a  href="#" id="btnJoin" class="button button-fill color-red toolbar-button">
                 加入团购
             </a>
