+++ front/html/group.html
@@ -28,7 +28,7 @@
                  </div>
             </div>
             <div class="card-footer">
-                <a href="#" class="link" onclick="myApp.alert('加入成功');">加入團購</a>
+                <a href="#" class="link" id="btnJoinInGroupPage">加入團購</a>
                 <a href="group-detail.html" class="link">查看詳情</a>
             </div>
         </div>
