+++ front/html/index.html
@@ -93,7 +93,7 @@
 <!-- Path to Framework7 Library JS-->
 <script type="text/javascript" src="../../bower_components/Framework7/dist/js/framework7.min.js"></script>
 <!-- Path to your app js-->
-<!--<script type="text/javascript" src="../dist/js/bundle.js"></script>-->
-<script type="text/javascript" src="../js/main.js"></script>
+<script type="text/javascript" src="../dist/js/bundle.js"></script>
+<!--<script type="text/javascript" src="../js/main.js"></script>-->
 </body>
 </html>
