+++ front/html/order.html
@@ -2,7 +2,7 @@
 <div data-page="order" class="page">
     <div class="navbar">
         <div class="navbar-inner">
-            <div class="left"><a href="index.html" class="back link icon-only"><i class="icon icon-back"></i></a></div>
+            <div class="left"><a href="group.html" class="back link icon-only"><i class="icon icon-back"></i></a></div>
             <div class="center">选择餐点·韩国紙上烤肉</div>
             <div class="right"><a href="#" class="open-panel link icon-only"><i class="icon icon-bars"></i></a></div>
         </div>
@@ -18,7 +18,7 @@
             <a class="button button-fill button color-green toolbar-button">
                 留言
             </a>
-            <a onclick="myApp.alert('跟團成功!')" href="index.html" class="button button-fill color-red toolbar-button">
+            <a onclick="" href="group.html" class="button button-fill color-red toolbar-button">
                 確認跟團
             </a>
         </div>
